backend/requirements.txt
psycopg2-binary
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-dotenv
authlib
httpx
google-generativeai
itsdangerous
